+++ python/lsst/ts/standardscripts/auxtel/__init__.py
@@ -24,6 +24,7 @@ from .enable_latiss import *
 from .latiss_take_sequence import *
 from .offline_atcs import *
 from .offline_latiss import *
+from .offset_atcs import *
 from .prepare_for.flats import *
 from .prepare_for.onsky import *
 from .shutdown import *