+++ python/lsst/ts/standardscripts/data/scripts/auxtel/offset_atcs.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 # This file is part of ts_standardscripts
 #
 # Developed for the LSST Telescope and Site Systems.
@@ -17,17 +18,10 @@
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .enable_comcam import *
-from .enable_mtcs import *
-from .offline_comcam import *
-from .offline_mtcs import *
-from .offset_mtcs import *
-from .setup_mtcs import *
-from .standby_comcam import *
-from .standby_mtcs import *
-from .stop import *
-from .take_image_comcam import *
-from .track_target import *
-from .track_target_and_take_image_comcam import *
-from .track_target_and_take_image_gencam import *
+import asyncio
+
+from lsst.ts.standardscripts.auxtel import OffsetATCS
+
+asyncio.run(OffsetATCS.amain())