+++ python/lsst/ts/standardscripts/data/scripts/maintel/laser_tracker/set_up.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 # This file is part of ts_standardscripts
 #
 # Developed for the LSST Telescope and Site Systems.
@@ -18,16 +19,8 @@
 #
 # You should have received a copy of the GNU General Public License
 
-from .enable_comcam import *
-from .enable_mtcs import *
-from .offline_comcam import *
-from .offline_mtcs import *
-from .offset_mtcs import *
-from .setup_mtcs import *
-from .standby_comcam import *
-from .standby_mtcs import *
-from .stop import *
-from .take_image_comcam import *
-from .track_target import *
-from .track_target_and_take_image_comcam import *
-from .track_target_and_take_image_gencam import *
+import asyncio
+
+from lsst.ts.standardscripts.maintel.laser_tracker import SetUp
+
+asyncio.run(SetUp.amain())