+++ python/lsst/ts/standardscripts/maintel/m1m3/__init__.py
@@ -18,16 +18,4 @@
 #
 # You should have received a copy of the GNU General Public License
 
-from .enable_comcam import *
-from .enable_mtcs import *
-from .offline_comcam import *
-from .offline_mtcs import *
-from .offset_mtcs import *
-from .setup_mtcs import *
-from .standby_comcam import *
-from .standby_mtcs import *
-from .stop import *
-from .take_image_comcam import *
-from .track_target import *
-from .track_target_and_take_image_comcam import *
-from .track_target_and_take_image_gencam import *
+from .raise_m1m3 import *